value: refactor out + add marshaling
This commit does a refactor of directory structure in anticipation of
adding packages for actionlog and profile, both of which need to use
Value and some other existing packages from engine.

This commit also installs some packages which are technically only needed for
future commits but it's painful to tease them apart, so I'm letting it be.