go: add instance abstraction; handle kafka setup
TODO: find a way of making tests faster

This commit introduces skeleton of an intstance abstraction that
can be used to selectively initialize / manage resources in prod
/ dev / test tiers.

Using the same abstraction, this commit also introduces some
initialization for kafka to make it work well in test settings
where we effectively want to start from scratch for every test.

As a result of this, we literally have to delete a kafka topic and reset
it again before every test. This slows down each test that depends on
Kafka to be 15sec. Further if multiple tests in a single test file
depend on kafka and/or http etc, we have to run them serially, which
further adds to test time. Currently, one of our end to end tests is
taking 30sec, which is very high. Ideally, we'd find some way of fixing
that.