kafka: make kafka its own package; client now doesn't talk to Kafka directly but via server
Earlier client was directly logging data to Kafka and for some calls it
was talking to the server. As a result, as we write more clients (e.g. Python)
we needed to maintain all state (e.g. kafka clusters/auth etc.) in Python
client as well, which is to be distributed to our customers. Instead, it is
much simpler if all the client has to do is talk to the server and then server
puts things in kafka or wherever else it needs to.