sql: replace default db engine in tests to mysql
The SQL dialect is meaningfully different between MySQL and SQLite
which required changing most of th queries.

Hence it's not possible to keep sqlite locally and mysql only
for remote. This commit starts using PlanetScale hosted mysql test
cluster for all unit tests.

While this does introduce dependency and slows down the tests a bit,
it seem to be fine for now. Eventually, if we are not happy with this
we may have to find a way of locally running or mocking MySQL