model: add a model for redis backed counters
In addition to being redis backed, this version is both simpler and more
powerful than the previous version. The main change is that we now
store everything in three granularities - minutes, hours, days. This way,
given any time period in question, we can approximate it with just days,
hours, and minutes (unlike before when we were dividing each window into
100 equal chunks). This approach simplifies the storage, will require
maintaining/updating fewer redis keys and is also more powerful because
we can now server arbitrary time windows.

TODO: get rid of the original model and rename this module to be just "counters"