interpreter: add basic benchmark to measure opcall performance
On a table of size 1K, when we do two opcall passes, it takes around 1.5ms right now.
This is super unoptimized so we should be able to squeeze in lots of performance. If
this performance stays with other parameters (e.g. more columns per table, more complex
opcalls etc.), we won't have to worry about performance of opcall evaluations.