e2e: Connect control-plane flask server to data-plane http-server
To run the app with a non-default end-point, run:
$ poetry run gunicorn 'app:build_app(endpoint="localhost:3000")' -b :2475

Since we run the flask server in gunicorn, we need to pass any runtime
flags as kwargs to the new build_app function. gunicorn does not allow
us to use regular command-line flags.

Currently, this is not done very well because of the flask `app` object
being a global and depending on a global client object `c`. We should
figure out a better way of doing this.