db: prefix each table with planeID
This now allows us to run multiple planes on the same database
(which corresponds to a tier). In addition to the abstraction,
this commit also updates all the existing models/tables to use
this abstraction and verifies all the tests are passing.

One nice side effect of this is that each unit test can get its
own plane within the same unit test tier.