otel: run open-telemetry collector as a single-replica deployment
This is a temporary work-around to prevent the same target from being
scraped by each replica and resulting in over-counting of metrics:
https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/main/receiver/prometheusreceiver/README.md

A better fix would be for us to run the otel collector as a sidecar and
have it only scrape local end-points. That way the collector will not
become a bottleneck.