kafka: improve integration testing system; add ctx to API
We now run kafka tests on mock kafka (normally) and with real kafka (in
integration mode). In addition, this commit improves the API to pass
context to kafka methods - unfortunately, this isn't natively supported
by the client library so we have to add a thin layer on top of it.