ingress: prefix chart name with namespace to remove conflicts
The emissary-ingress helm chart creates some cluster-level resources
(ClusterRole and ClusterRoleBinding) which create a conflict if we try
to create more than one ingress in the same cluster. We already know
that the helm chart name is prefixed to all resources, so we leverage
that to circumvent this problem by making the chart name specific to the
namespace in which we're trying to create the ingress.