eks: sleep after emissary-crds created before using CRD
This resolves the issue in provisioning the EKS project whereby creating
the l5d-mapping resource of type Mapping would fail because the service
emissary-apitext (in namespace emissary-system) wouldn't be ready by the
time the custom resource was being created.

Error:
Diagnostics:
  kubernetes:getambassador.io/v3alpha1:Module (l5d-mapping):
    error: resource default/ambassador was not successfully created by
    the Kubernetes API server : conversion webhook for
    getambassador.io/v3alpha1, Kind=Module failed: Post
    "https://emissary-apiext.emissary-system.svc:443/?timeout=30s": no
    endpoints available for service "emissary-apiext"

By adding a short 2 min sleep during creation time, we should circumvent
this problem.
The ideal solution here would have been to just run a "kubectl wait"
command that waits for service end-points to available/ready instead of
an arbitrary sleep, but unfortunately there seems to be no way to do
this as of k8s v1.21.