launchpad: create tier from plane creation output instead of env vars
We had to make a change to our pulumi project convention to allow this:
All `setup` functions in pulumi projects should not return an object of
type Promise<pulumi.Output<outputType>>, where the outputType itself
should not contain any pulumi.Ouput<> objects, since pulumi is able to
"lift" these types into an encapsulating Output type.

This change now makes it straightforward to return the output of
setupDataPlane as json from launchpad to the go control server.