eks: use kubectl wait for emissary-apiext deployment to be ready
This is needed for creating a Module customer-resource that can only be
satisfied by the emissary-apiext service. We had previously tried using
a 2 min sleep inside `emissaryCrds.resources.apply(...)` but
unfortunately pulumi does not guarantee that those resources have been
created before the sleep command is run.

Running a "kubectl wait" command with a long-enough timeout should be
sufficient to mitigate this issue.