db/schema: use if block of stored procedures to run schema updates
Earlier, 'alter table' was being run as part of a transaction. But DDL
statements (including alter table) don't play nice with transactions.
To fix that, we now run all statements conditionally (only when the
version is correct) - this way, each statement is run exactly once and
so we don't see any issues.