mock: fix use of mock in profile operator
Earlier mock ID was set in every case (due to the payload from client) which made
profile operator feel that it was in mock mode, even when it wasn't expected to be.

This was creating a bug. This commit fixes that bug. Someday, it will be good
to make the design of mock itself a bit more robust but this commit doesn't fix
that.