profile: add badger based models; reorganize tests
This commit adds support for profile based models - they are not used anywhere
right now though besides tests. As part of this, also reorganized unit tests
a bit to improve test code reuse.

This uncovered a bug as well - when we try to set a profile with same version to
different value, it doesn't work but gives an error. Ideal behavior will be
to given an error. This bug hasn't been fixed and affects both older DB/Cached
implementations. Badger implementation is correct wrt to this expectation