engine: compute signature only once for typechecking
Earlier we we were allocating a signature object for every single element
of each operator call, which cumulatively is a huge amount of allocation.

This used to take a massive amount of CPU earlier and also created
pressure on GC. This change fixes that by computing signatures only
once per operator call, not once for each element.