engine: stop doing list to dict conversion for ziptable
Earlier we were going to support Tuples and so didn't want to store
the heads as List. But now that is not the case and we ALWAYS
anyway store heads as a list of values. So better to store it as
List instead of Dict. In the latter, we do string allocation and
interpolation and later Atoi, for each element of each opcall. This
change should fix that and save us some allocations and CPU in the
most critical path of engine.