aggregate: don't double fetch duplicated aggregates
Fetching duplicated aggregate values is a common scenario (e.g. given a list
of contents, we try to fetch their topics, and then (uid, topic) based
aggregates which are duplicated across candidates. Earlier we weren't
deduping such aggregate calls. This commit fixes that. In addition, we
switch to using pre-allocated array lengths in the hope that the compiler
will learn to allocate them on the stack vs heap.