upgrade to go-1.18
This was required to setup capnproto go extension.
To workaround https://github.com/golang/go/issues/51991, we have
temporarily disabled tls for all connections to mysql and redis. Since
we will be moving away from Redis and RDS for storing any customer data,
this isn't a super-urgent problem for now and we can still claim to have
e2e encryption.