refac: separate histogram based value transform from bucketing
After this change, the Bucketize(...) function has a simpler job of
taking in a bucketizer and bucketizing the given actions as per the
bucketizer. It no longer concerns itself with transforming the given
actions to zero values of the given histogram - that responsibility is
moved to the MergeBuckets function which will be passed the output of
the Bucketize function.

With this change, the Bucketize function no longer needs to be tested
separately for different histogram implementations - the tests in
bucketizer_test.go are sufficient.