interpreter: replace a map with slice in innermost opcall loop
The innermost for loop in opcalls is when we evaluate each kwarg
in the context of each input. Previously, we used to set a map
from string to value. This commit changes that to a slice which is
lot more efficient both in writing as well as iteration/reading.

This should give us some CPU savings on engine.