interpreter: save a few fn calls & allocations in opcall hot path
A rather common scenario is that a kwarg is a constant (e.g. imagine
otype kwarg of profile being "user" and key kwarg being "age"). In
such cases, we don't need to set up lambda variables, which in turn
means that we don't need to create a function call. This should
hopefully give non-trivial CPU savings.