lib: add arena for allocating slices of generic types
It is hard to use sync.Pool for allocating arbitrary sized slices. If we
try to maintain lots of different pools, one for a particular size, the
overhead becomes way too high. This commit adds an arena which is still
concurrent and efficient (compared to allocating memory on heap in a loop),
but does by reserving a chunk of memory that is not released until the arena
is destroyed.

Very simple benchmark shows that it can be 3-4 faster than heap based allocation
when we can reuse the buffer again and again.