bucketizer: use arena in bucketize hotpath
This method of bucketizing durations is a memory hotpath on query servers.
This method starts using bucket arena to reduce allocations and memory
footprint.

To ensure that alloc/free are coordinated and don't go out of coordination even by
mistake in the future, this commit also moves the method that allocs in the same
struct that frees alloc (it was only used in one place anyway).