storage: use larger arena
Assuming 100 candidates per request and 300 buckets per candidate (we use 6
minute buckets in 2-level storage), we get 30K buckets for a single aggregate
call. If our arenas are smaller than that, we will not even take the arena
path and simply use heap allocated buckets. This commit increases the maxCap
for few arenas so that we start hitting them in our prod path.