storage: break down requests to redis in batches
Currently we go OOM because each batch of Redis blocks a lot of
RAM. If we have many concurrent redis requests (either due to high
QPS or some other issue like memorydb misbehaving), or the size of
request was large (e.g. because it had lots of candidates), a lot of
total RAM is blocked leading to OOM.

This commit fixes that by forcing a batch size of upto 16K buckets. This
should put a hard cap on how much RAM we can spend on a single request
(assuming GC / Arena is working well). The downside of this is on latency -
now we have to potentially make many more round trips to redis.