arena: export stats + start randomly dropping some Frees
Earlier the arena was fairly static - once a few large slices come in,
there isn't enough space left for smaller slices. And so we might get
lots of misses for smaller sizes because of a few large slices.

This commit repairs that to some degree by randomly dropping all free
calls with some probability. This "eviction" mechanism should help
bring the distribution of the arena more in sync with changing traffic.

In addition, this commit also exposes counters about the performance of
these arenas which will be useful in finetuning these over time.