counters: return value slice to arena using defer (#947)
Earlier, we would not just fail to return the slice back to the arena in
case of an error, we were simply not returning the slice back at all.

This was happening because we re-slice vals because of which it's
capacity only reflects left-over slots to the right. At the end, its len
and cap will both be zero and the slice will not be admitted into the
arena.