kafka inspector: stop busy reading/writing; use tickers
Earlier we were constantly reading a batch of messages, which can be
tens of thousands of messages every month, incurring a bill. This
commit starts doing this only in 5 minutely ticks. Also, this commit
starts recreating a client in each iteration which results in us always
reading the latest offset.