change aggregate durations to use uint32 throughout
This is related to changes in #925 and makes things more uniform while
we're using uint32 to represent second-granularity timestamps. If/when
we decide to move back to uint64, this change can be undone as well.

Note that changes to the proto types are also safe because the values
are variable-length encoded and therefore binary compatible:
https://developers.google.com/protocol-buffers/docs/proto#updating