counter: extract duration from kwargs in counter controller (#1103)
Instead of making the counter model kwarg aware, we extract the duration
in the controller and send that to the counter model functions. This
will allow us to re-use the same code path in nitrous as well as current
counter implemenation.

This commit also starts migration towards using aggtype constants
instead of magic strings that are spread throughout the code.