fix action controller test (#1122)
The TestKafkaInsertRead unittest would occasionally fail because writing
actions as JSON to actionlog_json topic is done asynchronously and it
is possible that actions could be re-ordered due to non-determinism in
go-routine scheduling. To fix the test, we only assert that elements
match instead of asserting that they are in the same order as input.