countaggr: disable readiness probe
We noticed that countaggr pods were stuck in a crash loop because the
readiness check was failing. While we don't yet understand why it was
happening, we temporary unblocked ourselves by disabling them. Note that
these are also not very relevant for countaggr since it doesn't accept
any service traffic.