rex-1518: fix bug in previous impl of ssh enablement
MangedNodeGroup looks for a ec2 keypair name to add it to instances on
launch. Earlier we used a keypair present in one region for the same
which would fail as we would use a different region for provisioning.
With this change we create a keypair on the fly if there don't exist one
already.