nitrous: remove CPU limit to avoid throttling
Adding a CPU limit can cause CPU throttling which can affect latency.
Since nitrous processes mostly run on dedicated machines, it is
acceptable to remove the CPU limit and allow the process to take as much
CPU as it needs. One downside of this change is that nitrous can starve
kubelet which can cause the node to go into a NotReady state.

Source:
https://erickhun.com/posts/kubernetes-faster-services-no-cpu-limits/