gravel: remove bloom, fix/skip a few tests
We could not deploy gravel earlier because race detection was flagging
an issue with Bloom Filters. We have now disabled it so that we can
start testing with disk hash solution

We could not deploy gravel earlier because race detection was flagging
an issue with Bloom Filters. We have now disabled it so that we can
start testing with disk hash solutoin