hangar/gravel: increase the pool size, make the queue even bigger
We have seen cases where the enqueue latencies are high even though CPU is available.
That's almost certainly because we don't have enough workers (especially since
we are now testing on a 64 CPU machine).

This commit makes this variable dependent on the number of cores. It also makes
the queue a bit larger because there is no point waiting on the queue - might
as well enqueue and then wait.