gravel: use power of 2 for bucket count
Looks like we are spending a lot of CPU on modulo operation. We can save
that by choosing a power of 2 as bucket count.

After this commit, future tables will have power of 2 bucket sizes but
reads will still work for older tables. Eventually, we should just get
rid of the codepath that uses modulo.