Nikhil/gravel hangar (#1421)
* utils/binary: add helpers to write non-varint integers in buffers

* gravel: checkin initial commit

This commit provides gravel - a fast key/value store based on hashmaps.
This is in "kind of working" state - table building is slow for large
datasets, lock management for memtables isn't robust, no metrics are
logged to prometheus yet and lots of other work still needs to be done.

* v0 of gravel/disk_hash_table

* bug fixes

* add benchmark for non-exist records

* gravel: remove bloom, fix/skip a few tests

We could not deploy gravel earlier because race detection was flagging
an issue with Bloom Filters. We have now disabled it so that we can
start testing with disk hash solution

We could not deploy gravel earlier because race detection was flagging
an issue with Bloom Filters. We have now disabled it so that we can
start testing with disk hash solutoin

* gravel: fix a bug + support deletion tombstone

Deletion tombstone is needed for hangar tests to pass

* gravel: clone data before putting in memtable

Earlier we were storing the same byte sequences created by callers of
gravel in memtable. But the callers could discard those buffers. We
now clone these before storing.

* hangar: add gravel based store

This passes all the tests of hangar interface.

* gravel: start exporting basic stats

* go: nitrous use gravel; fix lints; nitrous use new dir

* reduce counter contention and prefetch data index

* gravel: fix nitrous bump + add metrics

* gravel: remove layered cache

* gravel: Fix lint; msk: add volume size since we have hit the limits

* gravel: report sampled metrics correctly

This commit also starts using 'bitwise and' for sampling instead
of modulo. Probably not a huge deal but doing it since it's on
the hot path. Also, we now report metrics in a way that it's already
including the sampling effect.

* gravel: start recording the time taken by data reads

* hangar/gravel: increase the pool size, make the queue even bigger

We have seen cases where the enqueue latencies are high even though CPU is available.
That's almost certainly because we don't have enough workers (especially since
we are now testing on a 64 CPU machine).

This commit makes this variable dependent on the number of cores. It also makes
the queue a bit larger because there is no point waiting on the queue - might
as well enqueue and then wait.

* metrics: Increase resolution of the summary metrics and scrape nitrous metrics frequently

* hangar/gravel: don't report metrics for each key get

* gravel: close btree db before opening a read only copy

* gravel: flush memtable if it hasn't been flushed in 10 mins

* metrics: adjust the allowed error tolerance on quantiles

* nitrous: fix error message to refer to gravel, not pebble

* gravel: add a few error handling checks while writing files

This commits adds the following checks:
- While the file is being created, it uses a non-standard suffix which is
  renamed to the right suffix only as the last step. This way, if the temp
  file survives on the disk, it won't be read by the DB on next boot.

- A defer block is added which deletes even the temp file in case any error
  happens between creating the file and returning the values

- Adds an explicit sync call to minimize the probability of data not getting
  flushed to OS buffers

* gravel: throw explicit error if mmapped file is empty

* gravel: rename 'hits' metric to 'gets'

Hits as a name was a poor choice - we needed it to represent the total
number of requests, not the requests that succeed.

* gravel: don't time every data read

* utils/math: add function to get the next power of 2 for an integer

* gravel: use power of 2 for bucket count

Looks like we are spending a lot of CPU on modulo operation. We can save
that by choosing a power of 2 as bucket count.

After this commit, future tables will have power of 2 bucket sizes but
reads will still work for older tables. Eventually, we should just get
rid of the codepath that uses modulo.

* cleanup: remove trace file that was added by mistake

* gravel: fix a potential block condition in flush

The following scenario is possible - an organic flush happens and it
sends a message in the flushch. Meanwhile the periodicFlush goroutine
has just received a signal from the 10 min timer so it doesn't take
things out of the flushch and then later attempts to flush again.

* gravel: make flushch a buffered channel

This will guard against the case where 10min timer flush is trying to
flush - it won't have anyone else to pull from it. By making it
buffered, this can be made async.

* fix lints

* replace Tick with Ticker and wait on the channel

Co-authored-by: Siyuan Fu <siyuan@fennel.ai>
Co-authored-by: Mohit Reddy <mohitreddy1996@gmail.com>