nitrous: remove periodic flush which causes deadlock and stops binlog processing;
We essentially have 2 barriers - commitlock and flushch.

When only periodicflush is working - this took the commitLock and then waited indefinitely on flushCh, channel on which no can send a message because for that commitlock needs to be available.

When only SetMany is working (high write throughput) - this is fine since a non-blocking go routine is reading from flushch channel, so SetMany takes the lock, writes to flushch and writes the table.

Few Race condition -
1. 10 mins have passed and periodic flush attempts to take a lock. SetMany might call flush() and grab the lock first. Then try to write to flushch which will be blocked.
2. Otherway around, periodicFlush grabs the lock and then waits for flushch to which no one can write to because commitLock is grabbed.

nit