gravel: add sharding for tables; introduce manifest file
This commit shards Gravel storage. Each shard has its own lineage of
tables and its own memtable (though flushes of memtables for each
shard happen simultaneously).

To manage multiple shards and their files, this commit also introdueces
a manifest file which contains a list of all the files referenced
by each shard of the table. Manifest also remembers the number of
shards of the database so that subsequent openings can not mutate
data that way.