Nikhil/gravel shard (#1425)
* perf improvement

* nitrous: remove periodic flush which causes deadlock and stops binlog processing;

We essentially have 2 barriers - commitlock and flushch.

When only periodicflush is working - this took the commitLock and then waited indefinitely on flushCh, channel on which no can send a message because for that commitlock needs to be available.

When only SetMany is working (high write throughput) - this is fine since a non-blocking go routine is reading from flushch channel, so SetMany takes the lock, writes to flushch and writes the table.

Few Race condition -
1. 10 mins have passed and periodic flush attempts to take a lock. SetMany might call flush() and grab the lock first. Then try to write to flushch which will be blocked.
2. Otherway around, periodicFlush grabs the lock and then waits for flushch to which no one can write to because commitLock is grabbed.

nit

* nitrous: measure latency on the client side

* gravel: don't compute hash for each table again and again

This should save some CPU cycles, especially for databases with
many tables.

* cuckoo: add WIP version of cuckoo filter

Cuckoo filters are like bloom filters but usually better/faster and
can also be used to associate entries with keys. This implementation
is not complete yet - and may not even be needed for Gravel. If it is
not needed, I will remove it later.

* gravel: add sharding for tables; introduce manifest file

This commit shards Gravel storage. Each shard has its own lineage of
tables and its own memtable (though flushes of memtables for each
shard happen simultaneously).

To manage multiple shards and their files, this commit also introdueces
a manifest file which contains a list of all the files referenced
by each shard of the table. Manifest also remembers the number of
shards of the database so that subsequent openings can not mutate
data that way.

* gravel: use higher order 32 bits for computing fingerprint in diskhashtables

* gravel/hashtable: discard bits used for sharing from bucket calculation

* gravel/hashtable: remove the check for non-power of 2 bucket count

Earlier we were allowing non-power of 2 bucket counts and so were using modulo.
We have since then disallowed that. It wasn't changed earlier because
we had data in database. Now that we are purging everything, we can get
rid of this codepath too.

* cuckoo: remove code

It hasn't been used so far and is creating some unit test/lint issues.

* gravel: fix lint errors

* gravel: start reporting num tables

This commit also renames Lock/Unlock methods on manifest to
Reserve/Release to avoid any sort of interface collision issues
which may be causing the server to crash. Don't know if it work
or not - just trying out a random thing.

* nitrous: configure grpc streaming client to be aggressive while establishing new or broken connections

* linkerd: consider nitrous server's port as . This forces linkerd to work only at the TCP connection and does not care about the higher networking levels; This is to avoid linkerd's involvement in the connection latency observed, causing client to have requests piling up

* gravel/manifest: fix tests

* nitrous: set keepalive enforcement policy on nitrous server to respect keepalives from the client

* gravel: reimplement hashtable with new index + make code more readable

This commit adds a lot of documentation + breaks down GET/Table Build code
in smaller helpers, making it easier to read/update the code in the future.

As part of this, it also makes a few small changes to the index structure and
adds some tests + benchmarks.

* gravel: Fix lints in the test files

Co-authored-by: Siyuan Fu <siyuan@fennel.ai>
Co-authored-by: Mohit Reddy <mohitreddy1996@gmail.com>