kafka: reduce costly string allocation for each message
To check for repeated messages, we used to construct a string
representing each message. This allocation, subsequent GC and
operating on a map of strings all takes substantial CPU overhead
(as seen on Nitrous when doing 1M writes per second).

This change switches to a much lighter uint64 based message ID to
check for duplicates. This should significantly reduce the overhead
without affecting the correctness.