gravel: speed up table build by not sorting all the records
Earlier we were sorting all the records which was taking a LOT of
CPU. After this commit, we divide all the records in the appropriate
bucket and do sorting only within each bucket. On a table with 5M
records, the table build time went from 1.5sec -> 1sec for a 33%
improvement. I'd expect this improvement to be even larger for larger
tables.

This makes no functional change to the API.