gravel: speed up reads by using a single match index
Earlier we were populating an array of match indices. In CPU profiles,
this part was taking a lot of time (especially with defers). This also
meant that we had to scane the whole index line even  after getting a
key hit in the index. This was adding more CPU to readIndex
(which is called 8-10x more often, depending on the number of files in
a shard) to save CPU on readData side.

This commit changes readIndex to return a single match index if one
is found. This way, we don't need to pass around an array, don't need
a defer block, and we can also short circuit early if a key is found
=> this makes readIndex lot faster.

On readData side, if a fingerprint matches, the probability of key not
actually matching the key stored in data section is 1 / (2 ^ 24). So
by just matching a single index, we don't incur any slowdown in a vast
majority of cases.  As a result, this commit makes readIndex faster
for every single get while making readData slower for one in 1/(2^24)
calls which hit the data, which in turns is only 1/10th of all the calls.

Local benchmark shows that when keys are not in the table, the time taken
per key reduces by ~10% (from ~320ns to 290ns) for table of 5M keys. The
time taken to read a key in the table doesn't change much (because that
is presumably dominated by readData component)