gravel: don't create string maps during compaction
Earlier we were incurring a lot of CPU in GetAll of tables -
1) It was doing string allocations for each key in the table
2) It was adding each key in the map

Since GetAll returned a map, it had to one more iteration on it
to remove expired entries if doing lowest level compaction. All
of this was resulting in a very high CPU utilization of compaction.

This commit changes all of compaction path to only operate on slices
and also eliminates string allocations. This should hopefully reduce
both the compaction CPU and GC CPU (due to less memory garbage).

However, there are two downsides of this approach:
- Regular table build path now has one additional slice allocation to
  go from map to slice. This makes table builds slower/more CPU intensive
- Data of each table is mmaped, so instead of creating memory copy (via strings),
  we just return a pointer to this this mmaped portion. If the OS unmaps
  this segment, we could incur page cache thrash + make compaction MUCH
  slower. If it does happen, there are some ways of solving it.

Overall, I expect this to be net CPU friendly but we should observe this
on real datasets to be sure.