Siyuan/gravel 2bytes (#1490)
* nitrous prod read traffic replay changes required

* add a comment on the how the write and read are currently inconsistnet

* gravel: speed up table build by not sorting all the records

Earlier we were sorting all the records which was taking a LOT of
CPU. After this commit, we divide all the records in the appropriate
bucket and do sorting only within each bucket. On a table with 5M
records, the table build time went from 1.5sec -> 1sec for a 33%
improvement. I'd expect this improvement to be even larger for larger
tables.

This makes no functional change to the API.

* gravel/tests: use 5M keys in table test and report per key latencies

* gravel: speed up reads by using a single match index

Earlier we were populating an array of match indices. In CPU profiles,
this part was taking a lot of time (especially with defers). This also
meant that we had to scane the whole index line even  after getting a
key hit in the index. This was adding more CPU to readIndex
(which is called 8-10x more often, depending on the number of files in
a shard) to save CPU on readData side.

This commit changes readIndex to return a single match index if one
is found. This way, we don't need to pass around an array, don't need
a defer block, and we can also short circuit early if a key is found
=> this makes readIndex lot faster.

On readData side, if a fingerprint matches, the probability of key not
actually matching the key stored in data section is 1 / (2 ^ 24). So
by just matching a single index, we don't incur any slowdown in a vast
majority of cases.  As a result, this commit makes readIndex faster
for every single get while making readData slower for one in 1/(2^24)
calls which hit the data, which in turns is only 1/10th of all the calls.

Local benchmark shows that when keys are not in the table, the time taken
per key reduces by ~10% (from ~320ns to 290ns) for table of 5M keys. The
time taken to read a key in the table doesn't change much (because that
is presumably dominated by readData component)

* gravel: don't create string maps during compaction

Earlier we were incurring a lot of CPU in GetAll of tables -
1) It was doing string allocations for each key in the table
2) It was adding each key in the map

Since GetAll returned a map, it had to one more iteration on it
to remove expired entries if doing lowest level compaction. All
of this was resulting in a very high CPU utilization of compaction.

This commit changes all of compaction path to only operate on slices
and also eliminates string allocations. This should hopefully reduce
both the compaction CPU and GC CPU (due to less memory garbage).

However, there are two downsides of this approach:
- Regular table build path now has one additional slice allocation to
  go from map to slice. This makes table builds slower/more CPU intensive
- Data of each table is mmaped, so instead of creating memory copy (via strings),
  we just return a pointer to this this mmaped portion. If the OS unmaps
  this segment, we could incur page cache thrash + make compaction MUCH
  slower. If it does happen, there are some ways of solving it.

Overall, I expect this to be net CPU friendly but we should observe this
on real datasets to be sure.

* cpu optimization

* bug fix

* Revert "gravel: don't create string maps during compaction"

This reverts commit bb14a7782554265fce7f1a317cffe47ffa3ddf0c.

* add some metrics and revert a regression

* minor fixes

* bug fixes

* minor memory improvement

Co-authored-by: Mohit Reddy <mohitreddy1996@gmail.com>
Co-authored-by: Nikhil Garg <nikhilgarg28@gmail.com>
Co-authored-by: Siyuan Fu <siyuan@fennel.ai>