gravel: don't compute a separate hash for shard
In a local benchmark, I saw that significant CPU was being spent
on ShardHash calculation. This commit replaces that with some
bitwise manipulation on the main hash itself. Since lower 32 and
upper 32 bits are independent in any good hash, we derive a new
hash number out of it by xoring lower/upper bits with upper/lower
bits respectively.