gravel: switch to 32 byte index entries
Instead of trying to fit 19 fingerprints on average in 64 byte entries,
we now try to fit 9 fingerprints on average in 32 byte entries. This
still keeps the overhead at 3.5 bytes per record on average. But now
during gets, we need to scan only half the entries in both the index
and the data section. In a local benchmark, this increases the number
of gets per second by 30% or so.

Further, we can now remove the binary search in fps because the overhead
of binary search is not worth the smaller buffer scan anymore. Verified
locally that removing this makes no difference in the local benchmark.