add load limiter (#1507)
* gravel: flush the memtables before closing

This makes it less likely to "lose" updates. Making this change
to unblock a local benchmark I am doing.

* gravel: don't compute a separate hash for shard

In a local benchmark, I saw that significant CPU was being spent
on ShardHash calculation. This commit replaces that with some
bitwise manipulation on the main hash itself. Since lower 32 and
upper 32 bits are independent in any good hash, we derive a new
hash number out of it by xoring lower/upper bits with upper/lower
bits respectively.

* gravel: switch to 32 byte index entries

Instead of trying to fit 19 fingerprints on average in 64 byte entries,
we now try to fit 9 fingerprints on average in 32 byte entries. This
still keeps the overhead at 3.5 bytes per record on average. But now
during gets, we need to scan only half the entries in both the index
and the data section. In a local benchmark, this increases the number
of gets per second by 30% or so.

Further, we can now remove the binary search in fps because the overhead
of binary search is not worth the smaller buffer scan anymore. Verified
locally that removing this makes no difference in the local benchmark.

* add load limiter

* load limiter v2

* GC expired items more aggressively

* small fixes

* mem fixes

* minor fixes

* crash fix

Co-authored-by: Nikhil Garg <nikhilgarg28@gmail.com>
Co-authored-by: Siyuan Fu <siyuan@fennel.ai>