emissary: revert version to 8.5.1 (#1624)
The roll-out of 8.5.1 to Lokal's tier caused routing issues that we have
not yet been able to root-cause. Although the roll-out was first done to
our staging plane (t-108), that did not cause any issues, probably
because we only run one service (http-server).
To catch issues like this sooner the next time, this commit also adds
query-servers to the staging tier (t-108).