mothership: update kubeconfig and remove cert setup
The cert-manager was not able to satisfy the HTTP-01 challenge because
of acme-challenge not being routed correctly and receiving a 404 error.
We actually don't need to setup certs because we can just proxy the
request at our dns provider itself and use a self-signed certificate for
the communication between cloudflare and the backend load-balancer.